Reorganize the application section of the stake card
