+++ .env
@@ -1,4 +1,3 @@
-REACT_APP_SUPPORTED_CHAIN_ID=1337
-REACT_APP_ETH_HOSTNAME_HTTP=http://localhost:8545
-REACT_APP_ETH_HOSTNAME_WS=ws://localhost:8545
-REACT_APP_MULTICALL_ADDRESS=$MULTICALL_ADDRESS
+REACT_APP_SUPPORTED_CHAIN_ID=5
+REACT_APP_ETH_HOSTNAME_HTTP=https://goerli.infura.io/v3/33593948cb074eea8e65ae716fc61afd
+REACT_APP_ETH_HOSTNAME_WS=wss://goerli.infura.io/ws/v3/33593948cb074eea8e65ae716fc61afd
