+++ package.json
@@ -12,10 +12,10 @@
     "@ethersproject/constants": "^5.5.0",
     "@ethersproject/hardware-wallets": "^5.5.0",
     "@fontsource/inter": "^4.5.10",
-    "@keep-network/coverage-pools": "development",
-    "@keep-network/keep-core": "development",
-    "@keep-network/keep-ecdsa": "development",
-    "@keep-network/tbtc": "development",
+    "@keep-network/coverage-pools": "^1.1.0-goerli.0",
+    "@keep-network/keep-core": "1.8.1-goerli.0",
+    "@keep-network/keep-ecdsa": "^1.9.0-goerli.0",
+    "@keep-network/tbtc": "^1.1.2-goerli.0",
     "@ledgerhq/hw-app-eth": "^6.9.0",
     "@ledgerhq/hw-transport-u2f": "^5.36.0-deprecated",
     "@ledgerhq/hw-transport-webhid": "^6.7.0",
@@ -25,7 +25,7 @@
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
     "@threshold-network/components": "^1.0.0-dev.9",
-    "@threshold-network/solidity-contracts": "development",
+    "@threshold-network/solidity-contracts": "^1.2.0-goerli.0",
     "@types/ethjs-util": "^0.1.0",
     "@types/jest": "^27.0.1",
     "@types/ledgerhq__hw-transport-u2f": "^4.21.2",
